Rename: "파일 이동"
- vo 패키지를 만들어 vo에 해당되는 클래스들을 이동시킴.
