+++ src/test/java/baseball/model/ResultTest.java
@@ -1,5 +1,6 @@
 package baseball.model;
 
+import baseball.vo.Numbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
