+++ src/test/java/baseball/vo/NumbersTest.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
